separated time utility in its own file
