+++ include/net_setup.h
@@ -5,32 +5,21 @@
 #define LIGHTFX_NET_SETUP_H
 
 #include <WiFiNINA.h>
-#include <NTPClient.h>
 #include <Arduino_LSM6DSOX.h>
-#include <TimeLib.h>
-#include "secrets.h"
 #include "PaletteFactory.h"
 #include "util.h"
-
-#define CST_OFFSET_SECONDS (-21600)   //Central Standard Time - America/Chicago
-#define CDT_OFFSET_SECONDS  (-18000)  //Central Daylight Time - America/Chicago
+#include "secrets.h"
 
 extern WiFiServer server;
-
-void setupStateLED();
-void stateLED(CRGB color);
 
 bool wifi_setup();
 
 bool imu_setup();
-bool time_setup();
 void server_setup();
 void webserver();
 void wifi_loop();
 void printSuccessfulWifiStatus();
 void checkFirmwareVersion();
 uint8_t barSignalLevel(int32_t rssi);
-time_t curUnixTime();
-bool ntp_sync();
 
 #endif //LIGHTFX_NET_SETUP_H
