+++ src/fxB.cpp
@@ -62,7 +62,7 @@ void FxB::rainbow() {
 JsonObject & FxB1::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
 
@@ -138,7 +138,7 @@ void FxB::fxb_confetti() {
 JsonObject & FxB3::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
 
@@ -173,7 +173,7 @@ void FxB::sinelon() {
 JsonObject & FxB4::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
 
@@ -217,7 +217,7 @@ void FxB::juggle_short() {
 JsonObject & FxB5::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
 
@@ -285,7 +285,7 @@ void FxB::ease() {
 JsonObject & FxB7::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
 
@@ -330,7 +330,7 @@ void FxB::fadein() {
 JsonObject & FxB8::describeConfig(JsonArray &json) const {
     JsonObject obj = LedEffect::describeConfig(json);
     obj["brightness"] = brightness;
-    obj["palette"] = holidayToString(paletteFactory.currentHoliday());
+    obj["palette"] = holidayToString(paletteFactory.getHoliday());
     return obj;
 }
