+++ src/net_setup.cpp
@@ -5,11 +5,8 @@
 using namespace colTheme;
 const char ssid[] PROGMEM = WF_SSID;
 const char pass[] PROGMEM = WF_PSW;
-const char fmtDate[] PROGMEM = "%4d-%02d-%02d";
-const char fmtTime[] PROGMEM = "%02d:%02d:%02d";
 
 WiFiUDP Udp;  // A UDP instance to let us send and receive packets over UDP
-NTPClient timeClient(Udp, CST_OFFSET_SECONDS);  //time client, retrieves time from pool.ntp.org for CST
 
 /**
  * Convenience to translate into number of bars the WiFi signal strength received from {@code WiFi.RSSI()}
@@ -39,7 +36,7 @@ bool wifi_connect() {
     uint8_t wifiStatus = WiFi.status();
     while (wifiStatus != WL_CONNECTED) {
         if (attCount > 60)
-            stateLED(CRGB::Red);
+            updateStateLED(CRGB::Red);
         Log.infoln(F("Attempting to connect..."));
 
         // Connect to WPA/WPA2 network
@@ -83,7 +80,7 @@ bool imu_setup() {
     // initialize the IMU (Inertial Measurement Unit)
     if (!IMU.begin()) {
         Log.errorln(F("Failed to initialize IMU!"));
-        stateLED(CRGB::Red);
+        updateStateLED(CRGB::Red);
         //rtos::ThisThread::terminate();
         while (true) yield();
     }
@@ -91,91 +88,6 @@ bool imu_setup() {
     // print the board temperature
     boardTemperature();
     return true;
-}
-
-bool time_setup() {
-    //read the time
-    bool ntpTimeAvailable = ntp_sync();
-    setSyncProvider(curUnixTime);
-#ifndef DISABLE_LOGGING
-    Log.warningln(F("Acquiring NTP time, attempt %s"), ntpTimeAvailable ? "was successful" : "has FAILED, retrying later...");
-#endif
-    Holiday hday;
-    if (ntpTimeAvailable) {
-        bool bDST = isDST(timeClient.getEpochTime());
-        setSysStatus(SYS_STATUS_NTP);
-        if (bDST) {
-            setSysStatus(SYS_STATUS_DST);
-            timeClient.setTimeOffset(CDT_OFFSET_SECONDS);   //getEpochTime calls account for the offset
-        } else
-            timeClient.setTimeOffset(CST_OFFSET_SECONDS);   //getEpochTime calls account for the offset
-        setTime(timeClient.getEpochTime());    //ensure the offset change above (if it just transitioned) has taken effect
-        hday = paletteFactory.adjustHoliday();    //update the holiday for new time
-#ifndef DISABLE_LOGGING
-        Log.infoln(F("America/Chicago %s time, time offset set to %d s, current time %s. NTP sync ok."),
-                   bDST?"Daylight Savings":"Standard", bDST?CDT_OFFSET_SECONDS:CST_OFFSET_SECONDS, timeClient.getFormattedTime().c_str());
-        char timeBuf[20];
-        formatDateTime(timeBuf, now());
-        Log.infoln(F("Current time %s %s (holiday adjusted to %s"), timeBuf, bDST?"CDT":"CST", holidayToString(hday));
-#endif
-    } else {
-        resetSysStatus(SYS_STATUS_NTP);
-        hday = paletteFactory.adjustHoliday();    //update the holiday for new time
-        bool bDST = isDST(WiFi.getTime() + CST_OFFSET_SECONDS);     //borrowed from curUnixTime() - that is how DST flag is determined
-        if (bDST)
-            setSysStatus(SYS_STATUS_DST);
-#ifndef DISABLE_LOGGING
-        char timeBuf[20];
-        formatDateTime(timeBuf, now());
-        Log.warningln(F("NTP sync failed. Current time sourced from WiFi: %s %s (holiday adjusted to %s)"),
-              timeBuf, bDST?"CDT":"CST", holidayToString(hday));
-    }
-    Log.infoln(F("Current holiday is %s"), holidayToString(hday));
-#else
-    }
-#endif
-    return ntpTimeAvailable;
-}
-
-/**
- * Formats the time component of the timestamp, using a standard pattern - @see #fmtTime
- * @param buf buffer to write to. If not null, it must have space for 9 characters
- * @param time the time to format, if not specified defaults to @see now()
- * @return number of characters written to the buffer for given time value
- */
-uint8_t formatTime(char *buf, time_t time) {
-    if (time == 0)
-        time = now();
-    if (buf == nullptr)
-        return snprintf(buf, 0, fmtTime, hour(time), minute(time), second(time));
-    return snprintf(buf, 9, fmtTime, hour(time), minute(time), second(time));   //8 chars + null terminating
-}
-
-/**
- * Formats the date component of the timestamp, using a standard pattern - @see #fmtDate
- * @param buf buffer to write to. If not null, it must have space for 11 characters
- * @param time the time to format, if not specified defaults to @see now()
- * @return number of characters written to the buffer for given time value
- */
-uint8_t formatDate(char *buf, time_t time) {
-    if (time == 0)
-        time = now();
-    if (buf == nullptr)
-        return snprintf(buf, 0, fmtDate, year(time), month(time), day(time));
-    return snprintf(buf, 11, fmtDate, year(time), month(time), day(time));   //10 chars + null terminating
-}
-
-uint8_t formatDateTime(char *buf, time_t time) {
-    if (time == 0)
-        time = now();
-    uint8_t sz = formatDate(buf, time);
-    if (buf == nullptr)
-        sz += formatTime(buf, time);
-    else {
-        *(buf + sz) = ' ';  //date - time separation character
-        sz += formatTime( buf+sz+1, time);
-    }
-    return sz;
 }
 
 /**
@@ -209,7 +121,7 @@ bool wifi_check() {
  */
 void wifi_reconnect() {
     resetSysStatus(SYS_STATUS_WIFI);
-    stateLED(CRGB::Orange);
+    updateStateLED(CRGB::Orange);
     server.clearWriteError();
     WiFiClient client = server.available();
     if (client) client.stop();
@@ -218,7 +130,7 @@ void wifi_reconnect() {
     WiFi.end();     //without this, the re-connected wifi has closed socket clients
     delay(2000);    //let disconnect state settle
     if (wifi_connect())
-        stateLED(CRGB::Indigo);
+        updateStateLED(CRGB::Indigo);
     //NVIC_SystemReset();
 }
 
@@ -231,14 +143,14 @@ void wifi_loop() {
 
         if (!timeClient.isTimeSet()) {
             if (time_setup())
-                stateLED(CRGB::Indigo);
+                updateStateLED(CRGB::Indigo);
             else
-                stateLED(CRGB::Green);
+                updateStateLED(CRGB::Green);
         }
         Log.infoln(F("System status: %X"), getSysStatus());
     }
     EVERY_N_HOURS(12) {
-        Holiday oldHday = paletteFactory.currentHoliday();
+        Holiday oldHday = paletteFactory.getHoliday();
         Holiday hDay = paletteFactory.adjustHoliday();
 #ifndef DISABLE_LOGGING
         if (oldHday == hDay)
@@ -281,29 +193,3 @@ void checkFirmwareVersion() {
         Log.warningln(F("Please upgrade the WiFi firmware to %s"), WIFI_FIRMWARE_LATEST_VERSION);
     }
 }
-
-///////////////////////////////////////
-// Time sync
-///////////////////////////////////////
-
-time_t curUnixTime() {
-    if (timeClient.isTimeSet())
-        return timeClient.getEpochTime();
-    if (isSysStatus(SYS_STATUS_WIFI)) {
-        //the WiFi.getTime() (returns unsigned long, 0 for failure) can also achieve time telling purpose
-        //determine what offset to use
-        time_t wifiTime = WiFi.getTime() + CST_OFFSET_SECONDS;
-        if (isDST(wifiTime))
-            wifiTime = WiFi.getTime() + CDT_OFFSET_SECONDS;
-        return wifiTime;
-    }
-    return 0;
-}
-
-bool ntp_sync() {
-    timeClient.begin();
-    timeClient.update();
-    timeClient.end();
-    return timeClient.isTimeSet();
-}
-
